Correction
C7 should be 68p and DNP (do not populate).    Normally, for a SAR type input,  best results are obtain with an external resistor and capacitor to act as a charge reservoir for the sampling capacitor.    But, the analog inputs of microcontrollers used in Teensy, Arduino, SMT, etc., typically have a large resistance in series with the sampling capacitor.  This large resistance preempts any benefit from the external capacitor and instead lengthens the time constant for charging the sampling capacitor.   In this revision, we replace the external cap with one having a smaller value and we mark it as do-not-populate.   Indeed it should be left not populated.