Correct footprint for U4 (74LVC3G34)
With apologies, this update corrects the footprint for the 74LVC3G34, to match that for the part listed in the schematic BOM, the NXP 74LVC3G34DP,125, Digikey part number 1727-6685-1-ND.

The KiCAD footprint library contains two versions of TSSOP-8, size 4.4mm and 3.3m.  This part matches the smaller TSSOP-8.